remove dead code and move files
