+++ src/features/user-stats/components/stats-chart/stats-chart.component.tsx
@@ -1,4 +1,4 @@
-import { getColorFromCSSVar } from "shared/components/progress/utils"
+import { getColorFromCSSVar } from "features/home/utils"
 import React, { useMemo } from "react"
 import { Chart, AxisOptions, UserSerie } from "react-charts"
 import { CSSColorVariables } from "shared/constants"
