+++ src/shared/audio.d.ts
