--- src/shared/components/progress/Progress.module.css
@@ -1,8 +1,0 @@
-.progress-spiral {
-  max-width: 100%;
-  max-height: 100%;
-  aspect-ratio: 1;
-  display: block;
-  border-radius: 50%;
-  z-index: 1;
-}
