--- src/shared/components/progress/progress.constants.ts
@@ -1,8 +1,0 @@
-export const CANVAS_SIZE = 2520
-export const HALF_SIZE = CANVAS_SIZE / 2
-export const FIRST_POINT = { x: 0, y: 0 }
-export const SECOND_POINT = { x: -2.17, y: -2.17 } // rough value for nice rotate angle of spiral
-export const CENTER_POINT = {
-  x: HALF_SIZE,
-  y: HALF_SIZE,
-}
